+++ pong.go
@@ -11,6 +11,68 @@ import (
 // setting window variables
 const winWidth, winHeight int = 800, 600
 
+type gameState int
+
+const (
+	start gameState = iota
+	play
+)
+
+var state = start
+
+// fonts
+var nums = [][]byte{
+	{ // 0
+		1, 1, 1,
+		1, 0, 1,
+		1, 0, 1,
+		1, 0, 1,
+		1, 1, 1,
+	},
+	{ // 1
+		1, 1, 0,
+		0, 1, 0,
+		0, 1, 0,
+		0, 1, 0,
+		1, 1, 1,
+	},
+	{ // 2
+		1, 1, 1,
+		0, 0, 1,
+		1, 1, 1,
+		1, 0, 0,
+		1, 1, 1,
+	},
+	{ // 3
+		1, 1, 1,
+		0, 0, 1,
+		1, 1, 1,
+		0, 0, 1,
+		1, 1, 1,
+	},
+}
+
+func drawNumber(pos pos, color color, size, num int, pixels []byte) {
+	startX := int(pos.x) - (size*3)/2
+	startY := int(pos.y) - (size*5)/2
+
+	for index, v := range nums[num] {
+		if v == 1 {
+			for y := startY; y < startY+size; y++ {
+				for x := startX; x < startX+size; x++ {
+					setPixel(x, y, color, pixels)
+				}
+			}
+		}
+		startX += size
+		if (index+1)%3 == 0 {
+			startY += size
+			startX -= size * 3
+		}
+	}
+
+}
+
 // setting composite structs that will be used somewhere else
 type color struct {
 	r, g, b byte
@@ -49,7 +111,7 @@ func getCenter() pos {
 // computes new position, and collision logic
 func (ball *ball) update(leftPaddle, rightPaddle *paddle, elapsedTime float32) {
 	// log ball current speed
-	fmt.Printf("%+v %+v\n", *&ball.xv, *&ball.yv)
+	//fmt.Printf("%+v %+v\n", *&ball.xv, *&ball.yv)
 
 	// update ball position
 	ball.x += ball.xv * elapsedTime
@@ -60,21 +122,30 @@ func (ball *ball) update(leftPaddle, rightPaddle *paddle, elapsedTime float32) {
 	if ball.y-ball.radius < 0 || int(ball.y+ball.radius) > winHeight {
 		ball.yv *= -1
 	}
+
 	// if ball hits either left and right walls then reset position
-	if ball.x < 0 || int(ball.x) > winWidth {
+	if ball.x < 0 {
+		rightPaddle.score++
 		ball.pos = getCenter()
+		state = start
+	} else if int(ball.x) > winWidth {
+		leftPaddle.score++
+		ball.pos = getCenter()
+		state = start
 	}
 
 	// if the balls position is inside the left paddle
 	if ball.x-ball.radius < leftPaddle.x+leftPaddle.w/2 {
 		if ball.y > leftPaddle.y-leftPaddle.h/2 && ball.y < leftPaddle.y+leftPaddle.h/2 {
 			ball.xv *= -1
+			ball.x = leftPaddle.x + leftPaddle.w/2.0 + ball.radius
 		}
 	}
 	// same as above but for right paddle
 	if ball.x+ball.radius > rightPaddle.x-rightPaddle.w/2 {
 		if ball.y > rightPaddle.y-rightPaddle.h/2 && ball.y < rightPaddle.y+rightPaddle.h/2 {
 			ball.xv *= -1
+			ball.x = rightPaddle.x - rightPaddle.w/2.0 - ball.radius
 		}
 	}
 }
@@ -84,7 +155,13 @@ type paddle struct {
 	w     float32
 	h     float32
 	speed float32
+	score int
 	color color
+}
+
+// linear interpolation function to get a value at a percentage between 2 values
+func lerp(a float32, b float32, pct float32) float32 {
+	return a + pct*(b-a)
 }
 
 // drawing the paddle
@@ -98,11 +175,15 @@ func (paddle *paddle) draw(pixels []byte) {
 			setPixel(startX+x, currentY, paddle.color, pixels)
 		}
 	}
+
+	numX := lerp(paddle.x, getCenter().x, 0.2)
+	drawNumber(pos{numX, 35}, paddle.color, 10, paddle.score, pixels)
 }
 
 // updating the paddle, TODO: implement bounds for moving off the screen
 func (paddle *paddle) update(keyState []uint8, elapsedTime float32) {
-	fmt.Printf("paddle speed: %+v\n", *&paddle.speed)
+	// log current paddle speed
+	//fmt.Printf("paddle speed: %+v\n", *&paddle.speed)
 	// add or subtract the paddle velocity to the position to move the paddle
 	// if the paddles position is at the top of the window, it will no longer go up
 	// and the same for the bottom of the window,
@@ -177,8 +258,8 @@ func main() {
 
 	// initializing the game entities
 	// PLAYER INIT
-	player1 := paddle{pos{50, 300}, 20, 100, 300, color{255, 255, 255}}
-	player2 := paddle{pos{float32(winWidth) - 50, 300}, 20, 100, 300, color{255, 255, 255}}
+	player1 := paddle{pos{50, 300}, 20, 100, 300, 0, color{255, 255, 255}}
+	player2 := paddle{pos{float32(winWidth) - 50, 300}, 20, 100, 300, 0, color{255, 255, 255}}
 	ball := ball{getCenter(), 20, 200, 200, color{255, 255, 255}}
 
 	keyState := sdl.GetKeyboardState()
@@ -198,19 +279,28 @@ func main() {
 			}
 		}
 
-		// Update object positions
-		clear(pixels)
-
 		// Log current positions
 		// fmt.Println("Player 1 position:", player1.pos)
 		// fmt.Println("Player 2 position:", player2.pos)
 		// fmt.Println("Ball position:", ball.pos)
 
-		//updates
-		player1.update(keyState, elapsedTime)
-		player2.aiUpdate(&ball, elapsedTime)
-		ball.update(&player1, &player2, elapsedTime)
+		if state == play {
+			//updates
+			player1.update(keyState, elapsedTime)
+			player2.aiUpdate(&ball, elapsedTime)
+			ball.update(&player1, &player2, elapsedTime)
+		} else if state == start {
+			if keyState[sdl.SCANCODE_SPACE] != 0 {
+				if player1.score == 3 || player2.score == 3 {
+					player1.score = 0
+					player2.score = 0
+				}
+				state = play
+			}
+		}
 
+		// Update object positions
+		clear(pixels)
 		// Draw the objects
 		player1.draw(pixels)
 		player2.draw(pixels)
@@ -243,3 +333,5 @@ func main() {
 // i think the balls weird collision bug is related to the collisions of the ball being treated as a circle when it should be treated as a square
 
 // TODO: change ball collision to compute the balls physics as if it were a square instead of a ball
+// i added a fix that prevents the collision from happening with the paddle
+// but theres still a bug where the ball phases through the walls